Use city_io_to_geosjon repo as submodule
